feat: merge KeyValueTableWidget and KeyValueTableWidgetV2 (#688)
This is probably the smarter way of changing the constructor param types
rather than just having an entirely new class.

---

Example change to replace deprecated constructor usage:

```diff
new KeyValueTableWidget([
-  ["foo", "bar"],
+  { key: "foo", value: "bar" },
-  ["baz", "qux"],
+  { key: "baz", value: "qux" },
]);
```

---

Example change to replace deprecated V2 usage:

```diff
- new KeyValueTableWidgetV2([
+ new KeyValueTableWidget([
  { key: "foo", value: "bar" },
  { key: "baz", value: "qux" },
]);
```

---

_By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license_
