--- test/dashboard/widget/KeyValueTableWidgetV2.test.ts
@@ -1,11 +1,0 @@
-import { KeyValueTableWidgetV2, KeyValue } from "../../../lib/dashboard/widget";
-
-test("snapshot test", () => {
-  const data: KeyValue[] = [
-    { key: "name", value: "John Wick" },
-    { key: "has", value: "a dog" },
-  ];
-  const widget = new KeyValueTableWidgetV2(data);
-
-  expect(widget.toJson()).toMatchSnapshot();
-});
